[FIX] mrp_brewing: master_mo computation
master_mo computation only travelled to first parent and not
to the eldest MO in the creation hierarchy.
