+++ mrp_brewing/tests/__init__.py
@@ -1,1 +1,2 @@
 from . import test_brew_declaration
+from . import test_mrp_production
