+++ mrp_brewing/wizard/mrp_produce_wizard.py
@@ -9,15 +9,16 @@ class MRPProductProduce(models.TransientModel):
         mo_obj = self.env["mrp.production"]
 
         manufacturing_order = mo_obj.browse(self._context["active_id"])
-        # when does this happen ? The lot is created at all times
+        # previous lot may exist if produce wizard was closed
         product_lot = self.env["stock.production.lot"].search(
             [
                 ("name", "=", manufacturing_order.brew_order_name),
                 ("product_id", "=", manufacturing_order.product_id.id),
-            ]
+            ],
+            limit=1,
         )
 
-        if len(product_lot) > 0:
+        if product_lot:
             return product_lot.id
 
         vals = {
@@ -27,6 +28,4 @@ class MRPProductProduce(models.TransientModel):
         return self.env["stock.production.lot"].create(vals).id
 
     # Should only be visible when it is consume and produce mode
-    lot_id = fields.Many2one(
-        "stock.production.lot", "Lot", default=_create_production_lot
-    )
+    lot_id = fields.Many2one(default=_create_production_lot)
